Move ert_logging to ert.logging
