+++ .github/workflows/style.yml
@@ -47,7 +47,7 @@ jobs:
     - name: Run pylint
       run: |
         pip install .  # We need the dependencies of ERT to avoid import-error
-        pylint src/ert --ignore-paths=src/ert/gui
+        pylint src/ert --ignore-paths=src/ert/gui,src/ert/logging
         pylint --disable=all \
           --enable="\
             access-member-before-definition, \
@@ -107,4 +107,4 @@ jobs:
             useless-return, \
             useless-super-delegation, \
             " \
-          docs src/ert_data src/ert/gui src/ert_logging src/ert_shared src/job_runner src/res tests/*
+          docs src/ert_data src/ert/gui src/ert/logging src/ert_shared src/job_runner src/res tests/*
