+++ setup.py
@@ -48,8 +48,9 @@ args = dict(
     package_data={
         "ert_shared": package_files("src/ert_shared/share/"),
         "ert": package_files("src/ert/gui/resources/")
-        + package_files("src/ert/ert3/examples/"),
-        "ert_logging": ["logger.conf", "storage_log.conf"],
+        + package_files("src/ert/ert3/examples/")
+        + ["logging/logger.conf", "logging/storage_log.conf"],
+        "ert3_examples": package_files("src/ert3_examples/"),
         "res": [
             "fm/rms/rms_config.yml",
             "fm/ecl/ecl300_config.yml",
