+++ src/ert/logging/_log_util_abort.py
