+++ src/ert/logging/logger.conf
@@ -6,32 +6,32 @@ formatters:
     format: '%(asctime)s - %(name)s - %(threadName)s - %(levelname)s - %(message)s'
 handlers:
   file:
-    class: ert_logging.TimestampedFileHandler
+    class: ert.logging.TimestampedFileHandler
     level: DEBUG
     formatter: simple_with_threading
     filename: ert-log.txt
   experiment_server_file:
-    class: ert_logging.TimestampedFileHandler
+    class: ert.logging.TimestampedFileHandler
     level: DEBUG
     formatter: simple_with_threading
     filename: experiment-log.txt
   asyncio_file:
-    class: ert_logging.TimestampedFileHandler
+    class: ert.logging.TimestampedFileHandler
     level: DEBUG
     formatter: simple_with_threading
     filename: asyncio-log.txt
   event_log_file:
-    class: ert_logging.TimestampedFileHandler
+    class: ert.logging.TimestampedFileHandler
     level: DEBUG
     formatter: simple_with_threading
     filename: event-log.txt
   apifile:
-    class: ert_logging.TimestampedFileHandler
+    class: ert.logging.TimestampedFileHandler
     level: DEBUG
     formatter: simple
     filename: api-log.txt
   eefile:
-    class: ert_logging.TimestampedFileHandler
+    class: ert.logging.TimestampedFileHandler
     level: DEBUG
     formatter: simple_with_threading
     filename: ee-log.txt
