+++ src/ert/logging/storage_log.conf
@@ -18,7 +18,7 @@ handlers:
     stream: ext://sys.stdout
   file:
     formatter: standard
-    class: ert_logging.TimestampedFileHandler
+    class: ert.logging.TimestampedFileHandler
     level: DEBUG
     filename: api-log.txt
 loggers:
