+++ tests/ert_tests/shared/test_log_abort.py
@@ -1,6 +1,6 @@
 import logging
 from unittest.mock import MagicMock
-from ert_logging._log_util_abort import _log_util_abort
+from ert.logging._log_util_abort import _log_util_abort
 
 
 def test_log_util_abort(caplog, monkeypatch):
